Extract floating window controller
